Find Available Rooms (initial draft)
We want to send a request with the number of guests and bags required.

This method will search for rooms that are unoccupied and clean.

For this first pass lets only assign guests with bags to rooms that have storage and
guests with no bags to rooms that have no storage.

A third case (later) is if we have a guest without a bag that just needs a bed, in theory
we could find a room occupied that has a free bed and make that an option.  That would give
us maximum capacity.